requirements.txt
Flask==2.1.2
langchain==0.1.5
google-generativeai==1.0.0